Online Stuff (#6)
* Online evaluation/chatbox stuff

More lines for the chatbox output, with a smaller font(Total size is the same). I'm also submitting changes to _chatbox.lua in BAnims to properly change the size of the quad/square for the background of the chatbox.

Added [ScreenNetEvaluation] and added a second set of judgements and options which update when you scroll through users when you play online. Since it's in NetEvaluation it doesn't affect regular offline play.

* Add files via upload

Grade stuff for the done-through-metrics second set of judgements and grade.
And.
A couple BPM redirs for the BPm Display in the NetSelectMusic

* Create md

* Delete md

* Create asd.a

* Add files via upload

* Delete asd.a

* Create a

* Create a

* Create a

* Add files via upload

* Add files via upload

* Delete a

* Delete a

* Add files via upload

* Delete ScreenNetSelectMusic decorations.lua

* Delete ScreenNetSelectMusic underlay.redir

* Update metrics.ini

* Update ScreenNetEvaluation decorations.lua

* Update default.lua

* Now it uses Til Death's search system

* Add files via upload

* Delete currentsort.lua

* Delete currenttime.lua

* Delete default.lua

* Delete searchbar.lua

* Add files via upload

* Update default.lua

* Search query input is done through screenTextEntry

Previous method was also writing the query in the chatbox.

* Update default.lua