Completely changed userlist
Each username gets a font of it's own so we can change it's colour depending on state.
I think playerstate 1 is lobby, 2 is room, 3 is in options and 4 ingame.
Up to 8 users get a width of 50, if there are more than 8 their width gets reduced to 25. This is adjustable from 1 simple variable.