Enable Remapping RestartGameplay
Moved the key binding and so on from the theme to the game code. This allows for rebinding through the game options in a way that is less breakable/editable by the end user but still customizable to an extent.
The way to disable the button now would be to rebind the default key ( ` ) to something else like Coin instead of ripping out the theme code.