Skip sending judgment messages for misses in practice mode
This will heavily reduce stutter when skipping around.
The denser the song and the deeper the seek, the worse the lag. By skipping message sending, we can avoid most of the lag.
Tested on a 6k note song, the lag is reduced to less than 200ms rather than as much as 1000ms.