Refactored CPack setup to be isolated to single file
Much of the root CMakeLists.txt is only separated when necessary by operating system. This works for 
library linking. For CPack, since it's creates a distribution file based on previous commands, it's easier to have all those commands in the same place, so the full process of how and what is included in the binary is in one place.