Removed Windows from Travis-CI
Travis-CI is not at the point where we can use it for Etterna CI. Though they are close to it. If OpenSSL could be preinstalled on Windows VM's, it would be possible to completely switch over to travis for CI/CD. The OpenSSL.light package within chocolatey only contains the DLL files, not the header files which are also needed by Etterna.