Clean up .gitignore
The previous .gitignore was filled with a lot of lines which are useful only if the game was being built as an out-of-source build. While some  parts of the codebase are still used out-of-source, the majority of the builds are contained within the Build directory. That isolation allows much of the .gitignore to be removed.