Restore WindowsResources.rc to compilation process
This file was removed to evaulate it's necessity. It was a heavy-handed removal being reversed in this commit, as this file gives the game it's error popup and executable icon. The file is still reliant on an unnecessary smzip.ico, and likely could be simplified and moved to a more suitable location.