Implement ninja-build support
Since some libraries are statically linked, we had to add the logic to tell what to statically link if we're using ninja. Anything built from source does not require this logic. Maybe this can be revisited in the future to simplify it, since it is significant duplicate code, though I believe this is preferable to having a custom `USING_NINJA` variable in the CMake script.