Update libmad to include config file
The commit fixes a critical bug where all audio files would not be properly read, and therefore cause the game on startup when reading from the Songs directory. The config.mad.in.h was restored, and added into the CMakeLists.txt file to be generated into a config.h in the libmad build directory.