Fix some broken string compare logic in DLMAN
These comparisons have a 1/999999999999999999 chance of being true. The reason for that is because GetString returns a pointer and the hardcoded string it's being compared to isn't a pointer. So comparing these is comparing pointers. So we have to use std::strcmp. It returns 0 when a match occurs.
This issue broke some uncaught things along with online player rating updating.