Rebind Leaderboard loading to the wheel settling, kind of
it isn't really the wheel settling. this is a message triggered by checkbackgroundrequests but only happens to run whenever the wheel settles.
earlier recommendations were made to do this and now it is necessary since sample music does not natively play anymore if you have preview opened, and we still want leaderboards to load at any possible instance (as long as they are desired)