Fix mine bug
This is a bug where you can avoid pressing/holding a mine and then attempt to press a note that comes immediately after a mine. Doing so causes the game to try to score the mine, but gives you a miss because it is either not able to be judged or you are not required to step on mines directly.
This change makes the search used by Step() ignore mines that have been passed and just return the next upcoming tap (or mine)