Fix RecalculateSSRs race condition
this causes crashes and rare impossible scores on startup
at this point in song loading, usually a Steps's timingdata is empty. when you try to get the timingdata of a Steps in that state, it gives you the Song's timingdata. that's bad for when you rarely iterate through 2 highscores that are on different charts of the same song at the same time, because the Etaner of the timingdata gets cleared in one instance while messing with it in another function (we pass references around instead of copies, so messing with the Etaner is bad)