Fix Blind modifier causing notes to cease existing
Blind should only cause your gameplay to be "blind" as in "you have no reliable feedback"
So only Judgments and Combo should be invisible. In addition, the color of the explosions, if applicable in the Noteskin, should be constantly Marvelous even on a tap release so feedback is even more useless.