Band-aid Replay issue allowing Player to tap any note that exists
basically setting this to -1 allows the Step function to tap anything it wants since the score we give that Step is according to the ReplayData. So if it fails for some reason, it scores a possibly incorrect and possibly otherwise impossible tap. This causes Replays to look very bad. A regular example was found and this solves that. The other example is itsjake's 1h54m js 1.4x. That case is still not solved, but plays back much better using this change.