Fix macOS Catalina Crashing
macOS Catalina had a change in their OpenGL API. Specifically one which prevents the game from running if mac related UI updates are done off of the main thread. The issue is fixed by running the line which crashes the application in an asynchronous block to be sent to the main thread for execution.