Update most things to let the game compile again
MAJOR slowdown to startup due to inefficiencies in handling the difficultyrating struct. especially abusing skillset_vector which will generate a new vector to be used most often. this is not particularly useful in the cases where it is used. it would probably be better to reformat the difficultyrating struct to use a vector instead or write some fancy code to make constructing an entire new vector unnecessary