Add Service Locator and implementation for IFileManager
This commit will introduce a ServiceLocator. It will be used as a decoupling strategy, and implemented more thoroughly in future branch to isolate the globals all into a single class. While this branch will focus on rewriting the file manager, it will also convert the usage of that global this service locator setup.