Move ArchHooks from global to Service Locator
This is the commit where global `HOOKS` is removed from the source code. ArchHook ownership is moved to Locator.cpp which handles the lifetime of the object and bringing the class to other places in the source.

It's important to note that most of the uses of HOOKS is a single use in a lot of the imported files. Due to it's sparse use in the majority of the codebase, it could be possible to narrow the scope of where ArchHooks is imported.