Implement support for Debug and MinSizeRel
CMakeWindows.cmake was not setup properly for a multi-configuration generator like MSVC. Directly checking CMAKE_BUILD_TYPE is only sufficient for single-config generators. For multi-config, the build config (debug, release, etc) is not known until attempting to compile in the application, so generator expressions must be used to change how the cmake runs for each build config.

Other changes include:
- Giving each output binary it's own name so "Etterna.exe" is not overwritten each time
- Condense `set_target_properties` in root CMakeLists.txt. All the options are unnecessary considering what the build script currently supports.