Better fix for macOS black screen bug, fixes to incorrect std::chrono math
jameskr97's commit fixed the macOS black screen issue, but did so incorrectly.
This commit fixes the issue in a better manner, making it more maintainable for the future
The prior commit mistakenly worked by setting the second value of abstime to the microsecond count of pTimeout->c_dur, an incredibly huge number. (Around 80 billion, converting to around 2 thousand years)
Because this number was large enough to be effectively infinite, this commit accidentally solved the main issue, which was that the pthread_cond_timedwait call had always been timing out because the value of pTimeout being passed to it was far too small.

This commits accomplishes two things:
1) Fixing the incorrect std::chrono math being performed.
   This is done by setting abstime.tv_sec to the actual second count with a std::chrono::duration_cast, and by using timeofday in the else branch, instead of pTimeout.
2) Directly replacing the pthread_cond_timedwait call with a pthread_cond_wait call.
   This directly bypasses the timeout value abstime, which is currently a garbage value.
   Doing this, along with the comment, makes it explicitly clear that this is only a stopgap measure,and that abstime is not being used.

However, this commit doesn't fix the root issue, which is that EventImpl_Pthreads::Wait is being passed garbage input.