Fix Xcode build placing binary in incorrect folder
Since Xcode is also a multi-configuration generator, we have to tell Xcode where to place the binary, otherwise, it will always be placed in a folder. In order to remove that folder, we put an "always-true" CMake generator expression. That is documented to prevent the automatic insertion of a folder for each binary. The same can be seen in CMakeWindows.cmake to force every binary to be build in the Programs directory.