Move pack download curl handling into its own thread
uhh so is there any really good justification for this? no
anyone doing this responsibly would use curl easy perform since its in its own thread
on top of that it introduces a lot of possible race conditions i totally didnt fix by just putting a mutex around some of the things which definitely doesnt do anything im like 6% positive
is this a bad idea? maybe like half honestly
everything still works
this led to the discovery that differentialreload cannot be called from a thread separate from the main thread as d3d explodes and kills the game (deadlock) if you do that