Make Highest-Only mode work correctly in filters
Earlier, highest-only mode would allow songs that had a particular skillset
as the highest in any of its charts to be included, even if the chart
that did so did not meet other filter criteria.

This commit changes the IsSkillsetHighestOfAnySteps function into a new
IsSkillsetHighestOfChart function, which takes in a Steps* as a parameter.