Fix broken texture IDs in Banner/BG loaded Textures
also return self in the bg/banner lua hooks for chaining
this also includes all of the ragetexture pointer to shared pointer changes
so basically what the problem is, is that the RageTextureIDs of two equivalent textures is coming out slightly different - when creating a texture the colorDepth of the new ID is -1 so a new texture is always created, but logic fails somewhere down the line causing references to start leaking all over the place
this should actually be the single change that fixes most of the RageTextureManager leaks.
It introduces huge i/o related lag if you are heavily using Sprite:LoadBanner or Sprite:LoadBackground, which makes sense to me because you are literally loading an image from disk and projecting it into the game.