Switch static std::array to std::vector for interval based MinaCalc containers
Due to a change in MSVC 16.7 it appears that static arrays of a large size take a simply unacceptable amount of RAM/heap to compile. This kills GHA as well as almost our local machines. This is a change that "fixes" the issue by removing the very large static arrays.

Although this behavior may be due to a real MSVC bug, this change is useful on its own to allow a somewhat greater range of files. But still, the difference between 40k intervals and 100k intervals for a max size is basically nothing in the grand scheme of files available. Most files fall within 500-1000 intervals (4-8 minutes).
#890