Make Song::HighestMSDOfSkillset only match charts passing filter, if set
Earlier, a bug was caused where "Overall" and other skillset sorts would be
incorrect, as the sort would find the highest MSD chart of *any* from the
current game mode, even those hidden by the filter. This lead to confusing
behavior.

This fix simply replaces the GetChartsOfCurrentGameMode call with a
GetChartsMatchingFilter. (Which falls through to the above if the filter is
not set)