change folder lamp calculation, disregard sub C grades, add Clear lamps
lamps for a folder are the lowest common highest grade per file
to expand on the definition, require that all grades are C or better before counting, and also consider 1.0+ and downrate scores separately. if the highest rate cleared on a file is below 1.0, the highest grade for the file is a ghost grade called Grade_Tier20, which we will refer to as the Clear grade.
the reason for the C requirement is that F is equivalent to D in most respects, especially if you turn on FailOff
if the highest grade for a file is Grade_Tier20, the pack gets a Clear flag as long as all other files in the pack have any passing score (C or better)