Avoid stack overflow caused by ChartMatchesFilter calling HighestMSDOfSkillset which called ChartMatchesFilter itself
This was caused by a previous change which made HighestMSDOfSkillset only consider filter-matching charts, in order to fix sorting by skillset difficulties.
This had the side effect of breaking the filter option of highest difficulty only, which also used the same function.

This commit fixes the issue by adding in a boolean flag to HighestMSDOfSkillset, which toggles between considering filtered charts and charts of the current game mode.