NoteFieldPreview steps loaded from lua will follow style track remapping
it seems this only matters for beat which has a scratch lane and for some odd reason we decided to put column 8 on the first column, shifting everything right by 1. this implies that all single7 charts are actually for p2 which seems really really wrong or maybe our notesloader is wrong. didnt check that.
anyways this is proper behavior, whatever is going on with that is probably not