Frame timing with high precision timers
RageTimer was calculating a tight bound to wait, now() + frameTime,
but its calculation of now() rounded down to the millisecond. Then, it
waited with microsecond precision. This underestimates the correct
time to wait with jitter that averages out to the millisecond, which
resulted in the 500/1000fps binning for frame limits in that range.

This commit also moves all frame timing nonsense off the v-sync
and frame-limit-less paths.