Tune frame timing
Took some measurements this time. Now, frame time corrections are
on the order of microseconds and robust to frame skips. Seems to react
instantly to changes in steady-state clock speed.

By the way, another way to implement this is to sync to a clock of our
own we increment every frame by the target frame time (plus some
handling for skips). But it could alias with the monitor vsync, right?