Fix bug with non-exclusive filters and required Length/Clear% ranges
Both Length and Clear% must always be matched, even in non-exlusive mode.
However, this has the effect of non-exlusive mode just becoming any chart
that matches either of those ranges, regardless of skillsets.

This changes them to require a skillset to be matched if any are set; if
no skillset ranges are set, a non-exclusive filter using only length/clear%
behaves identically to an exclusive one.