pull jack diffs from anchorseq logic but leave anchorseq intact | 453
introduce oop to the structs in genericsequencing and upend the entire deal of how this stuff is working and make it much more confusing
this could be done better even still

but basically what this commit does is duplicate all the anchor logic and separate into one specific for anchors and one specific for jacks. this sounds counterproductive because anchors are jacks. but we want to handle jacks for jack difficulty in a special way that might not necessarily fit the anchor sequencer. so to leave the anchor sequencer intact while doing this, we separate the two while making them initially 2 of the same thing here in the first place

it would probably be better to just break this like the ohjsequencer or whatever and have params for it but this is what i did and its what im going with until it doesnt work

version iteration because i undid the jack diff parameter changes that were done in the past couple of calc iterations, but only to the anchor sequencer so that whatever really specific nuance they served for tech diff can remain while we work on jacks for jack diff instead of anchors as a broad subject