Tighten up hit offset computation
Player::Step has to translate a wall-clock time to a music position, and
it has GAMESTATE->m_Position as a point of reference, a point in time
where both wall and music time are known.

But instead of computing the difference between the hit time and the
wall time directly, the previous code was doing everything relative to
::now(). Mathematically this is alright, but it was calling ::now()
twice to do this, and the OS is free to schedule us out between those
two calls. If that happens, the note offset computation is just wrong.
Unlikely, sure, but it's easy to just be correct here.

Fixing this makes the code simpler too, so that's nice.