replace grindscaler with experimental length-weight grind scaler | 461
crammed it into nps::actual_cancer because thats what this is
needs a little tuning maybe. also softened the old grindscaler curve and used it in this. made the actual density weights very very light because this can easily destroy the rating for a file and since it is interval based this can flip flop up and down