Fix multithreaded bug in RageSound
This should fix https://crash.etterna.dev/crash-reports/2af49c0d-c8a0-4fe7-932b-592cdd9c6947
The decode thread could block in RageSound::GetDataToPlay waitin for
the recent samples mutex, and then the music thread in GameSoundManager
could delete the RageSound, which also locks the mutex.
If the deleting thread locked it first, and the decode thread waited for
it, it deleted the fft Plan and then the decode thread tried to use it.
Calling Unload() before deleting the plan should guarantee that the
decode thread no longer tries to use that RageSound. Unload() locks g_Mutex
which the GameSoundManager thread locks before deleting the RageSound.