Fix LuaHelpers::TypeError bug
To trigger it call a C++ function from lua that expects an Actor, and it must not be in the 'self' position. An easy to test example is self:SetFakeParent(1), which will erroneously say the argument is a boolean instead of a number. With this patch, it correctly identifies it as a number