The Rebirth theme is the default, not Til' Death.
In this commit, I fixed mentions of Til' Death being the default. The Rebirth theme is now the default for Etterna, not Til' Death. I capitalized the "e" in Etterna in the path to the en.ini file and put it inside a `` block, and added a space before the parentheses on the last sentence of the paragraph.