Allow for use of new XCode Build System without errors
The new XCode build system doesn't allow the same file to be generated by two independent targets without
 having it be a dependency of some common ancestor first.

This could be worked around by using the old XCode build system with a special CMake flag, but this produces
 some rather spooky warnings about the fact that it will be removed at some indeterminate point in the future.

Both the cURL and LuaJIT CMakes violate this requirement.

This commit fixes both very simply --
For cURL, we add in a third random target to be a parent to both that generates the common files, fixing the issue.
For LuaJIT, we do the same, although the parent target already exists (containing only the shared file
 complained about in the first place)