"improve" logic for post-search selection & resolve past wheeldata
the old logic here relied on the fact that the musicwheel leaks a list of pointers every time you song search and they sit in space forever. this fixes that by making sure that if those pointers were destroyed, we can find a replacement. if no replacement is found, ...ahh whatever

also changed search logic just a little so that it tries to stick you on your last selected song in more cases