Make Til Death determine BPM range based on specific chart rather than song
Previously, Til Death called the Lua binding for BPMDisplay::SetBpmFromSong
as BPMDisplay::SetBpmFromSteps was broken. This meant that songs that had
different BPM ranges per-chart (as is possible in .ssc files) only showed
the BPM range of the very first chart for all of them.

This small change fixes that and lets Til Death display the correct
ranges for every individual chart :)