replay rewrite wip: a lot
the majority of this commit contains code that never runs and is effectively dead. permanently dead in fact. its code that lets us test whether or not converting a replay from v2 to input or from input to v2 works successfully. and any code changes that arent within those two functions are extra changes that were the result of finding issues from using those functions to test things.
so as of now, we have replays about 99% safely converting in both directions with a wife percent error threshold of 0.1%. the majority of the errors are on the order of 0.0001% and are due to misses being reordered. aka: float precision errors.

what's left is to allow reprioritizing note judgments using inputdata. i think.

i also learned that all input data recorded up to this point needs to be invalidated for a lot of reasons. so, the read and write functions for it have been modified appropriately here to cause that to occur.