(Again) disable use of libssh2 when building cURL
Without this change, we run into the problem where our GHA builder has the
libssh2 library installed (and so the built cURL dynamically links to it) but
the average mac user does not, leading to the application immediately
crashing on start.

This was already fixed once before, but with the recent(ish) upgrade to a
newer version of cURL than what we'd been using in f1dd6fe, the option
CMAKE_USE_LIBSSH2 was renamed to CURL_USE_LIBSSH2, which silently made
our forcing of that setting to OFF no longer work.