GetDirListing actually can filter dirs and files; fix exts breaking song load
this fixes the crashing that occurs when you put a random audio file in a pack folder. this also fixes the crashes that occur if you name a song folder with a regular file extension. this also fixes errors that occur when naming a song folder as an image. and everything in between kind of. these song folders should load properly if they are real song folders. if you happen to do it in a particularly wrong way and song load ends prematurely you also get a very obvious warning in the log