break replay playback event generation to fix replay playback
yeah really so basically what i did here was do the right thing and it caused the player to automatically recalculate all the values based on current conditions. this fixed one specific issue that is caused when you record a replay on one specific song/global offset and then change either one and then play it back. checking back other replay types and regular input data replays this didnt break those so it looks fine. this just means half the code i wrote there to generate the stuff is kind of pointless