get ready to start organizing packs and tags
we find packs by doing a specific search, which may or may not include tags. we cache the packs that are returned, and have to create a pagination object for these unique calls. frontend will use the pagination objects to go through the returned data in a quicker way instead of repeating the api request when going backwards in pages