Create a version of Etterna that works with OSX High Sierra
The version of OpenSSL the CI uses targets too recent a macOS version,
resulting in it referencing the symbol ___chkstik_darwin, a stack-smashing
protection function added to macOS's libSystem after High Sierra.

This chang tells the GHA CI to build a version of Etterna (denoted Legacy)
that references a custom Homebrew tap I created instead of the default
OpenSSL homebrew tap.

My tap just adds in the option when building OpenSSL that tells it to
target an early macOS version :)